data_ingestion/requirements/base.txt
fastapi
uvicorn
pandas
sqlalchemy
python-dotenv
unidecode
cloudscraper
selenium
undetected_chromedriver
packaging
lxml
psycopg2-binary